be/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.27
boto3==1.20.8
flask_cors==3.0.10
flask_jwt_extended==4.3.1
awscli==1.22.8